WS-2759: Refactors most read to extract function
