+++ src/app/components/MostRead/Canonical/Rank/index.tsx
@@ -1,10 +1,3 @@
-import {
-  Burmese,
-  Bengali,
-  EasternArabic,
-  Nepali,
-  WesternArabic,
-} from '../../../../legacy/psammead/psammead-locales/src/numerals';
 import { Services } from '../../../../models/types/global';
 import { ColumnLayout, MostReadRankProps, Size } from '../../types';
 import styles, {
@@ -13,17 +6,7 @@ import styles, {
   getMultiColumnCss,
 } from './index.styles';
 
-export const serviceNumerals = (service: Services) => {
-  const servicesNonWesternNumerals = {
-    bengali: Bengali,
-    burmese: Burmese,
-    dari: EasternArabic,
-    nepali: Nepali,
-    pashto: EasternArabic,
-    persian: EasternArabic,
-  };
-  return servicesNonWesternNumerals[service] || WesternArabic;
-};
+import serviceNumerals from '../../utilities/getServiceNumerals';
 
 interface ColumnCssProps {
   listIndex: number | string;
